Refine ARM proxy machine client
