+++ pkg/aksmachine/client.go
@@ -8,6 +8,7 @@ import (
 
 func newMachineClientFromConfig(cfg *config.Config, logger *slog.Logger) (MachineClient, error) {
 	if cfg.Agent.ARMProxyURLOverrideForE2E != "" {
+		logger.Warn("using ARM proxy machine client for dev-test")
 		return newARMProxyClient(cfg, logger)
 	}
 	return newARMClient(cfg, logger)
